yield typed metric in DistributionAnalysisComponent
